refactor(core): decouple KgTripleRepository from SQLite (Phase 3)
